Added some initial code for creating audit records.
