+++ src/main/java/logging/AuditHelper.java
@@ -25,6 +25,7 @@ public class AuditHelper {
      */
     public AuditHelper() {
         gson = new GsonBuilder().create();
+        serviceName = System.getenv("SERVICE_NAME");
     }
 
     /**
@@ -41,5 +42,49 @@ public class AuditHelper {
         System.out.println("Audit JSON to be sent to GLADOS: " + outJson);
         //TODO implement API connection!
         return 501; //Unimplemented.
+    }
+}
+
+/**
+ * Audit Object class used with GSON.
+ */
+class AuditObj {
+
+    private String content;
+    private String serviceName;
+    private Date timestamp;
+    private String userId;
+
+    public AuditObj(String content, String serviceName, Date timestamp, String userId) {
+        this.content = content;
+        this.serviceName = serviceName;
+        this.timestamp = timestamp;
+        this.userId = userId;
+    }
+
+    public String getContent() {
+        return content;
+    }
+
+    public String getServiceName() {
+        return serviceName;
+    }
+
+    public Date getTimestamp() {
+        return timestamp;
+    }
+
+    public String getUserId() {
+        return userId;
+    }
+
+    @Override
+    public String toString() {
+        return "AuditObj{" +
+                "content='" + content + '\'' +
+                ", serviceName='" + serviceName + '\'' +
+                ", timestamp=" + timestamp +
+                ", userId='" + userId + '\'' +
+                '}';
     }
 }
